Add files via upload
the use of local variable and global variable.
Variables that are declared inside a function or block are local variables. They 
can be used only by statements that are inside that function or block of code. 
Local variables are not known to functions outside their own. Following is the 
example using local variables

Global Variables
Global variables are defined outside of all the functions, usually on top of the 
program. The global variables will hold their value throughout the life-time of 
your program.
A global variable can be accessed by any function. That is, a global variable is 
available for use throughout your entire program after its declaration. Following 
is the example using global and local variables